Refactor template constants
