+++ packages/edit-site/src/components/left-sidebar/navigation-panel/constants.js
@@ -1,0 +1,10 @@
+export const TEMPLATES_GENERAL = [
+	'front-page',
+	'archive',
+	'singular',
+	'index',
+	'search',
+	'404',
+];
+
+export const TEMPLATES_POSTS = [ 'home', 'single' ];
