+++ packages/edit-site/src/utils/get-template-info/constants.js
@@ -48,14 +48,3 @@ export const TEMPLATES_DEFAULT_DETAILS = {
 		description: __( 'Template for single posts' ),
 	},
 };
-
-export const TEMPLATES_GENERAL = [
-	'front-page',
-	'archive',
-	'singular',
-	'index',
-	'search',
-	'404',
-];
-
-export const TEMPLATES_POSTS = [ 'home', 'single' ];
